Move getting ext nodes to a different place
